+++ src/Chunkyard/ContentType.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Linq;
+
+namespace Chunkyard
+{
+    /// <summary>
+    /// Describes the different types a piece of content can be.
+    /// </summary>
+    public enum ContentType
+    {
+        Blob = 0,
+        Document = 1
+    }
+}