+++ src/Chunkyard.Build/Program.cs
@@ -17,7 +17,7 @@ namespace Chunkyard.Build
             }
             catch (Exception e)
             {
-                Console.WriteLine(e.Message);
+                Console.WriteLine($"Error: {e.Message}");
                 Environment.ExitCode = 1;
             }
         }